requirements.txt
Flask==3.0.2
Pillow==11.2.1
Werkzeug==3.0.1 